requirements.txt
boto3==1.17.46
botocore==1.20.46
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
marshmallow==3.11.1
numpy==1.20.2
pandas==1.2.3
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
SQLAlchemy==1.4.6
urllib3==1.26.4
Werkzeug==1.0.1
